Ensure that integration tests can run w.o. setup
The App-service not allows a config root as a parameter to the create
mehtod. This makes it possible to feed the config paths to when running
integration tests which is nice, because it doesn't need to go looking
in /etc/whatever.

Futhermore the port can now be picked up from the config, though sending
it from the commandline will override that. This is convinent since we
generally want to use supervisord to provide ports to the app - but this
way we can get the best of both worlds.